apis updates

+++ domain/model/grade.go
@@ -4,9 +4,9 @@ import "gorm.io/gorm"
 
 type Grade struct {
 	gorm.Model
-	Quiz  uint `json:"quiz"`
-	Score int  `json:"score"`
-	User  uint `json:"user"`
+	Quiz  uint    `json:"quiz"`
+	Score float64 `json:"score"`
+	User  uint    `json:"user"`
 }
 
 func (g *Grade) PublicGrade() *Grade {
